Modularize CloudSync runtime tests
Separate runtime tests into focused lifecycle, orchestration, status, shutdown, result, and native-interrupt modules without changing production behavior.
