+++ crates/db-core/src/cloudsync/runtime.rs
@@ -746,1676 +746,7 @@ fn embedded_sync_error(result: &CloudsyncNetworkResult) -> Option<String> {
 }
 
 #[cfg(test)]
-#[allow(clippy::items_after_test_module)]
-mod tests {
-    use super::*;
-
-    #[cfg(any(target_os = "macos", target_os = "linux", target_os = "windows"))]
-    async fn assert_interrupts_stalled_native_request_once() {
-        let listener = std::net::TcpListener::bind("127.0.0.1:0").unwrap();
-        let endpoint = format!("http://{}", listener.local_addr().unwrap());
-        let (accepted_tx, accepted_rx) = std::sync::mpsc::sync_channel(1);
-        let (release_tx, release_rx) = std::sync::mpsc::sync_channel(1);
-        let server = std::thread::spawn(move || {
-            let (stream, _) = listener.accept().unwrap();
-            accepted_tx.send(()).unwrap();
-            let _stream = stream;
-            let _ = release_rx.recv_timeout(Duration::from_secs(5));
-        });
-
-        let db = Arc::new(Db::connect_memory().await.unwrap());
-        sqlx::query(
-            "CREATE TABLE items (
-                id TEXT PRIMARY KEY NOT NULL,
-                value TEXT NOT NULL DEFAULT ''
-            )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.cloudsync_init("items", None, None).await.unwrap();
-        {
-            let mut connection = db.cloudsync_connection.lock().await;
-            *connection = Some(db.pool.acquire().await.unwrap());
-            sqlx::query("SELECT cloudsync_network_init_custom(?, ?)")
-                .bind(endpoint)
-                .bind("interrupt-test")
-                .fetch_optional(&mut **connection.as_mut().unwrap())
-                .await
-                .unwrap();
-        }
-
-        let request_db = Arc::clone(&db);
-        let request = tokio::spawn(async move {
-            let _sync_operation = request_db.cloudsync_sync_operation.lock().await;
-            let mut connection = request_db.cloudsync_connection.lock().await;
-            super::super::ops::interruptible_network_receive_changes(
-                connection.as_mut().unwrap(),
-                &request_db.cloudsync_interrupt,
-            )
-            .await
-        });
-        tokio::task::spawn_blocking(move || {
-            accepted_rx
-                .recv_timeout(Duration::from_secs(5))
-                .expect("native CloudSync request did not reach the blackhole server");
-        })
-        .await
-        .unwrap();
-
-        let interrupted_at = std::time::Instant::now();
-        while !request.is_finished() {
-            db.cloudsync_interrupt_sync();
-            assert!(
-                interrupted_at.elapsed() < Duration::from_secs(2),
-                "native CloudSync request did not honor sqlite3_interrupt"
-            );
-            tokio::time::sleep(Duration::from_millis(25)).await;
-        }
-        let result = request.await.unwrap();
-        assert!(result.is_err(), "interrupted native request succeeded");
-        assert!(interrupted_at.elapsed() < Duration::from_secs(2));
-        tokio::time::timeout(
-            Duration::from_millis(100),
-            db.cloudsync_wait_for_sync_idle(),
-        )
-        .await
-        .expect("sync operation released before its SQLite worker became idle");
-
-        {
-            let mut connection = db.cloudsync_connection.lock().await;
-            let value: i64 = sqlx::query_scalar("SELECT 1")
-                .fetch_one(&mut **connection.as_mut().unwrap())
-                .await
-                .unwrap();
-            assert_eq!(value, 1);
-            let worker_idle = connection.as_mut().unwrap().lock_handle().await.unwrap();
-            drop(worker_idle);
-        }
-
-        let _ = release_tx.send(());
-        server.join().unwrap();
-        db.cloudsync_close_connection().await.unwrap();
-    }
-
-    #[cfg(any(target_os = "macos", target_os = "linux", target_os = "windows"))]
-    #[tokio::test]
-    async fn native_cloudsync_interrupt_drains_blackhole_http_and_reuses_connection() {
-        for _ in 0..3 {
-            assert_interrupts_stalled_native_request_once().await;
-        }
-    }
-
-    fn test_cloudsync_config() -> CloudsyncRuntimeConfig {
-        CloudsyncRuntimeConfig {
-            connection_string: "sqlitecloud://demo.invalid/app.db?apikey=demo".to_string(),
-            auth: super::super::CloudsyncAuth::None,
-            tables: Vec::new(),
-            sync_interval_ms: 30_000,
-            wait_ms: Some(500),
-            max_retries: Some(1),
-        }
-    }
-
-    async fn db_with_local_unsent_changes() -> (tempfile::TempDir, Db) {
-        let dir = tempfile::tempdir().unwrap();
-        let db_path = dir.path().join("app.db");
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(2),
-        })
-        .await
-        .unwrap();
-        sqlx::query(
-            "CREATE TABLE items (
-                id TEXT PRIMARY KEY NOT NULL,
-                value TEXT NOT NULL DEFAULT ''
-            )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.cloudsync_init_enabled_tables(&[CloudsyncTableSpec {
-            table_name: "items".to_string(),
-            crdt_algo: None,
-            init_flags: None,
-            enabled: true,
-        }])
-        .await
-        .unwrap();
-        let mut connection = db.pool().acquire().await.unwrap();
-        sqlx::query("INSERT INTO items (id, value) VALUES ('item', 'pending')")
-            .execute(&mut *connection)
-            .await
-            .unwrap();
-        connection.return_to_pool().await;
-        assert!(
-            db.pool().num_idle() > 0,
-            "cloudsync test pool has no idle connections (size={}, max={})",
-            db.pool().size(),
-            db.pool().options().get_max_connections(),
-        );
-        (dir, db)
-    }
-
-    use std::future::pending;
-    use std::sync::atomic::{AtomicBool, AtomicUsize, Ordering};
-
-    use crate::{CloudsyncAuth, CloudsyncTableSpec, DbOpenOptions, DbStorage};
-    #[test]
-    fn embedded_sync_failures_update_runtime_error_state() {
-        let runtime = Mutex::new(CloudsyncRuntimeState::default());
-        runtime.lock().unwrap().last_sync_at_ms = Some(42);
-        let result = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "failed".to_string(),
-                local_version: 4,
-                server_version: 3,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 0,
-                tables: Vec::new(),
-                chunks: 0,
-                bytes: 0,
-                complete: true,
-                error: Some("schema mismatch".to_string()),
-                last_failure: None,
-            }),
-        };
-
-        record_sync_result(&runtime, result, false);
-
-        let runtime = runtime.lock().unwrap();
-        assert!(runtime.last_sync.is_some());
-        assert_eq!(runtime.last_sync_at_ms, Some(42));
-        assert_eq!(runtime.consecutive_failures, 1);
-        assert_eq!(
-            runtime.last_error_kind,
-            Some(anlg_cloudsync::ErrorKind::Fatal)
-        );
-        assert!(
-            runtime
-                .last_error
-                .as_deref()
-                .unwrap()
-                .contains("schema mismatch")
-        );
-    }
-
-    #[test]
-    fn embedded_sqlite_contention_remains_retryable() {
-        let runtime = Mutex::new(CloudsyncRuntimeState::default());
-        let result = CloudsyncNetworkResult {
-            send: None,
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 0,
-                tables: Vec::new(),
-                chunks: 0,
-                bytes: 0,
-                complete: false,
-                error: Some("database is locked".to_string()),
-                last_failure: None,
-            }),
-        };
-
-        record_sync_result(&runtime, result, false);
-
-        let runtime = runtime.lock().unwrap();
-        assert_eq!(
-            runtime.last_error_kind,
-            Some(anlg_cloudsync::ErrorKind::Transient)
-        );
-        assert_eq!(runtime.consecutive_failures, 1);
-    }
-
-    #[test]
-    fn embedded_sync_in_progress_preserves_last_successful_sync() {
-        let runtime = Mutex::new(CloudsyncRuntimeState::default());
-        runtime.lock().unwrap().last_sync_at_ms = Some(42);
-        let result = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "syncing".to_string(),
-                local_version: 4,
-                server_version: 3,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 3,
-                tables: vec!["sessions".to_string()],
-                chunks: 1,
-                bytes: 2048,
-                complete: false,
-                error: None,
-                last_failure: None,
-            }),
-        };
-
-        record_sync_result(&runtime, result, false);
-
-        let runtime = runtime.lock().unwrap();
-        assert_eq!(runtime.last_sync_at_ms, Some(42));
-        assert!(runtime.last_error.is_none());
-        assert_eq!(runtime.consecutive_failures, 0);
-    }
-
-    #[test]
-    fn initial_sync_stays_unsettled_while_receive_is_in_progress() {
-        let runtime = Mutex::new(CloudsyncRuntimeState::default());
-        let result = CloudsyncNetworkResult {
-            send: None,
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 3,
-                tables: vec!["sessions".to_string()],
-                chunks: 1,
-                bytes: 2048,
-                complete: false,
-                error: None,
-                last_failure: None,
-            }),
-        };
-
-        record_sync_result(&runtime, result, false);
-
-        let runtime = runtime.lock().unwrap();
-        assert!(runtime.last_sync_at_ms.is_none());
-        assert!(runtime.last_error.is_none());
-    }
-
-    #[test]
-    fn completed_receive_marks_sync_complete_without_a_send_result() {
-        let runtime = Mutex::new(CloudsyncRuntimeState::default());
-        let result = CloudsyncNetworkResult {
-            send: None,
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 0,
-                tables: Vec::new(),
-                chunks: 0,
-                bytes: 0,
-                complete: true,
-                error: None,
-                last_failure: None,
-            }),
-        };
-
-        record_sync_result(&runtime, result, false);
-
-        assert!(runtime.lock().unwrap().last_sync_at_ms.is_some());
-    }
-
-    #[test]
-    fn settled_network_preserves_last_success_while_local_work_remains() {
-        let runtime = Mutex::new(CloudsyncRuntimeState {
-            last_sync_at_ms: Some(42),
-            ..Default::default()
-        });
-        let result = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "synced".to_string(),
-                local_version: 2,
-                server_version: 2,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 0,
-                tables: Vec::new(),
-                chunks: 0,
-                bytes: 0,
-                complete: true,
-                error: None,
-                last_failure: None,
-            }),
-        };
-
-        record_sync_result(&runtime, result, true);
-
-        assert_eq!(runtime.lock().unwrap().last_sync_at_ms, Some(42));
-    }
-
-    #[test]
-    fn bounded_sync_combines_send_and_receive_results() {
-        let send = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "synced".to_string(),
-                local_version: 4,
-                server_version: 4,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: None,
-        };
-        let receive = CloudsyncNetworkResult {
-            send: None,
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 3,
-                tables: vec!["sessions".to_string()],
-                chunks: 1,
-                bytes: 2048,
-                complete: false,
-                error: None,
-                last_failure: None,
-            }),
-        };
-
-        let result = merge_bounded_sync_results(send.clone(), receive.clone());
-
-        assert_eq!(result.send, send.send);
-        assert_eq!(result.receive, receive.receive);
-        assert!(sync_result_needs_receive_progress(&result));
-    }
-
-    #[test]
-    fn background_sync_starts_immediately_and_continues_incomplete_receive_promptly() {
-        let interval = Duration::from_secs(30);
-        let incomplete = CloudsyncNetworkResult {
-            send: None,
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 1,
-                tables: vec!["e2ee_records".to_string()],
-                chunks: 1,
-                bytes: 1024,
-                complete: false,
-                error: None,
-                last_failure: None,
-            }),
-        };
-
-        assert_eq!(cloudsync_next_delay(None, false, interval), Duration::ZERO);
-        assert_eq!(
-            cloudsync_next_delay(Some(&incomplete), false, interval),
-            CLOUDSYNC_PROGRESS_INTERVAL
-        );
-
-        let send_in_progress = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "syncing".to_string(),
-                local_version: 3,
-                server_version: 2,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 0,
-                tables: Vec::new(),
-                chunks: 0,
-                bytes: 0,
-                complete: true,
-                error: None,
-                last_failure: None,
-            }),
-        };
-        assert_eq!(
-            cloudsync_next_delay(Some(&send_in_progress), false, interval),
-            CLOUDSYNC_PROGRESS_INTERVAL
-        );
-
-        let settled = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "synced".to_string(),
-                local_version: 3,
-                server_version: 3,
-                chunks: 0,
-                bytes: 0,
-                last_failure: None,
-            }),
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 0,
-                tables: Vec::new(),
-                chunks: 0,
-                bytes: 0,
-                complete: true,
-                error: None,
-                last_failure: None,
-            }),
-        };
-        assert_eq!(
-            cloudsync_next_delay(Some(&settled), true, interval),
-            CLOUDSYNC_PROGRESS_INTERVAL
-        );
-
-        let uploaded = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "synced".to_string(),
-                local_version: 4,
-                server_version: 4,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: settled.receive,
-        };
-        assert_eq!(
-            cloudsync_next_delay(Some(&uploaded), true, interval),
-            interval
-        );
-    }
-
-    #[derive(Default)]
-    struct RecordingSyncHook {
-        directive: crate::CloudsyncSyncDirective,
-        local_work_remaining: bool,
-        activity_paused: AtomicBool,
-        before_calls: AtomicUsize,
-        after_result: Mutex<Option<CloudsyncNetworkResult>>,
-    }
-
-    impl crate::CloudsyncSyncHook for RecordingSyncHook {
-        fn activity_paused(&self) -> bool {
-            self.activity_paused.load(Ordering::SeqCst)
-        }
-
-        fn before_sync<'a>(
-            &'a self,
-            _pool: &'a SqlitePool,
-        ) -> crate::CloudsyncBeforeHookFuture<'a> {
-            Box::pin(async move {
-                self.before_calls.fetch_add(1, Ordering::SeqCst);
-                Ok(self.directive)
-            })
-        }
-
-        fn after_sync<'a>(
-            &'a self,
-            _pool: &'a SqlitePool,
-            result: &'a CloudsyncNetworkResult,
-        ) -> crate::CloudsyncHookFuture<'a> {
-            Box::pin(async move {
-                *self.after_result.lock().unwrap() = Some(result.clone());
-                Ok(crate::CloudsyncHookOutcome {
-                    local_work_remaining: self.local_work_remaining,
-                    deferred: false,
-                })
-            })
-        }
-    }
-
-    #[tokio::test]
-    async fn before_sync_hook_can_select_receive_only_transport() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        let recording_hook = Arc::new(RecordingSyncHook {
-            directive: crate::CloudsyncSyncDirective::ReceiveOnly,
-            ..Default::default()
-        });
-        let hook: Arc<dyn crate::CloudsyncSyncHook> = recording_hook;
-        let hook = Mutex::new(Some(hook));
-
-        assert_eq!(
-            run_before_sync_hook(&hook, db.pool()).await.unwrap(),
-            crate::CloudsyncSyncDirective::ReceiveOnly
-        );
-    }
-
-    #[tokio::test]
-    async fn deferred_before_sync_hook_never_starts_native_transport() {
-        let db = Db::connect_memory().await.unwrap();
-        sqlx::query("CREATE TABLE items (id TEXT PRIMARY KEY NOT NULL)")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        db.cloudsync_init("items", None, None).await.unwrap();
-        let recording_hook = Arc::new(RecordingSyncHook {
-            directive: crate::CloudsyncSyncDirective::Deferred,
-            ..Default::default()
-        });
-        let hook: Arc<dyn crate::CloudsyncSyncHook> = recording_hook.clone();
-        let hook = Mutex::new(Some(hook));
-
-        let result = sync_cloudsync_connection(
-            db.pool(),
-            &db.cloudsync_connection,
-            &db.cloudsync_interrupt,
-            &db.cloudsync_sync_operation,
-            &db.cloudsync_runtime,
-            &hook,
-        )
-        .await
-        .unwrap();
-
-        assert!(matches!(result, CloudsyncStepOutcome::Deferred));
-        assert_eq!(recording_hook.before_calls.load(Ordering::SeqCst), 1);
-        assert!(recording_hook.after_result.lock().unwrap().is_none());
-    }
-
-    #[tokio::test]
-    async fn existing_native_pending_batch_skips_before_sync_hook() {
-        let db = Db::connect_memory().await.unwrap();
-        sqlx::query(
-            "CREATE TABLE items (
-                id TEXT PRIMARY KEY NOT NULL,
-                value TEXT NOT NULL DEFAULT ''
-            )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.cloudsync_init("items", None, None).await.unwrap();
-        sqlx::query("INSERT INTO items (id, value) VALUES ('item', 'pending')")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let recording_hook = Arc::new(RecordingSyncHook::default());
-        let hook: Arc<dyn crate::CloudsyncSyncHook> = recording_hook.clone();
-        let hook = Mutex::new(Some(hook));
-
-        let result = sync_cloudsync_connection(
-            db.pool(),
-            &db.cloudsync_connection,
-            &db.cloudsync_interrupt,
-            &db.cloudsync_sync_operation,
-            &db.cloudsync_runtime,
-            &hook,
-        )
-        .await;
-        let Err(error) = result else {
-            panic!("sync unexpectedly succeeded");
-        };
-
-        assert_eq!(recording_hook.before_calls.load(Ordering::SeqCst), 0);
-        assert!(
-            error
-                .to_string()
-                .contains("Unable to retrieve CloudSync network context")
-        );
-    }
-
-    #[tokio::test]
-    async fn activity_pause_precedes_an_existing_native_pending_batch() {
-        let db = Db::connect_memory().await.unwrap();
-        sqlx::query(
-            "CREATE TABLE items (
-                id TEXT PRIMARY KEY NOT NULL,
-                value TEXT NOT NULL DEFAULT ''
-            )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.cloudsync_init("items", None, None).await.unwrap();
-        sqlx::query("INSERT INTO items (id, value) VALUES ('item', 'pending')")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let pending_before = {
-            let mut connection = db.pool().acquire().await.unwrap();
-            crate::cloudsync::ops::ensure_pending_payload_fits(
-                &mut connection,
-                &db.cloudsync_interrupt,
-            )
-            .await
-            .unwrap()
-        };
-        assert!(pending_before.chunks > 0);
-
-        let recording_hook = Arc::new(RecordingSyncHook {
-            activity_paused: AtomicBool::new(true),
-            ..Default::default()
-        });
-        let hook: Arc<dyn crate::CloudsyncSyncHook> = recording_hook.clone();
-        let hook = Mutex::new(Some(hook));
-        let last_sync = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "synced".to_string(),
-                local_version: 4,
-                server_version: 4,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: None,
-        };
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.last_sync = Some(last_sync.clone());
-            runtime.last_sync_at_ms = Some(42);
-            runtime.last_error = Some("previous error".to_string());
-        }
-
-        let result = sync_cloudsync_connection(
-            db.pool(),
-            &db.cloudsync_connection,
-            &db.cloudsync_interrupt,
-            &db.cloudsync_sync_operation,
-            &db.cloudsync_runtime,
-            &hook,
-        )
-        .await;
-
-        assert!(matches!(result, Ok(CloudsyncStepOutcome::Deferred)));
-        assert_eq!(recording_hook.before_calls.load(Ordering::SeqCst), 0);
-        assert!(recording_hook.after_result.lock().unwrap().is_none());
-        let pending_after = {
-            let mut connection = db.pool().acquire().await.unwrap();
-            crate::cloudsync::ops::ensure_pending_payload_fits(
-                &mut connection,
-                &db.cloudsync_interrupt,
-            )
-            .await
-            .unwrap()
-        };
-        assert_eq!(pending_after, pending_before);
-        let runtime = db.cloudsync_runtime.lock().unwrap();
-        assert_eq!(runtime.last_sync.as_ref(), Some(&last_sync));
-        assert_eq!(runtime.last_sync_at_ms, Some(42));
-        assert_eq!(runtime.last_error.as_deref(), Some("previous error"));
-    }
-
-    #[cfg(any(target_os = "macos", target_os = "linux", target_os = "windows"))]
-    #[tokio::test]
-    async fn activity_pause_during_pending_preflight_defers_and_drains_before_local_write() {
-        let db = Arc::new(Db::connect_memory().await.unwrap());
-        sqlx::query(
-            "CREATE TABLE items (
-                id TEXT PRIMARY KEY NOT NULL,
-                value TEXT NOT NULL DEFAULT ''
-            )",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-        db.cloudsync_init("items", None, None).await.unwrap();
-        sqlx::query("SELECT cloudsync_set('payload_max_chunk_size', '33554432')")
-            .fetch_optional(db.pool())
-            .await
-            .unwrap();
-        sqlx::query(
-            "WITH RECURSIVE rows(id) AS (
-                SELECT 1
-                UNION ALL
-                SELECT id + 1 FROM rows WHERE id < 100000
-            )
-            INSERT INTO items (id, value)
-            SELECT printf('item-%d', id), printf('value-%d', id) FROM rows",
-        )
-        .execute(db.pool())
-        .await
-        .unwrap();
-
-        let recording_hook = Arc::new(RecordingSyncHook::default());
-        db.set_cloudsync_sync_hook(recording_hook.clone());
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.running = true;
-            runtime.network_initialized = true;
-        }
-
-        let request_db = Arc::clone(&db);
-        let request = tokio::spawn(async move { request_db.cloudsync_trigger_sync().await });
-        let started_at = std::time::Instant::now();
-        while !db.cloudsync_interrupt_registered() {
-            assert!(
-                !request.is_finished(),
-                "pending payload generation completed before interrupt registration was observable"
-            );
-            assert!(
-                started_at.elapsed() < Duration::from_secs(2),
-                "pending payload generation never registered for interruption"
-            );
-            tokio::task::yield_now().await;
-        }
-
-        recording_hook.activity_paused.store(true, Ordering::SeqCst);
-        while !request.is_finished() {
-            db.cloudsync_interrupt_sync();
-            assert!(
-                started_at.elapsed() < Duration::from_secs(2),
-                "pending payload generation did not honor sqlite3_interrupt"
-            );
-            tokio::task::yield_now().await;
-        }
-
-        assert_eq!(
-            request.await.unwrap().unwrap(),
-            CloudsyncNetworkResult::default()
-        );
-        assert!(db.cloudsync_runtime.lock().unwrap().last_error.is_none());
-        assert_eq!(recording_hook.before_calls.load(Ordering::SeqCst), 0);
-        assert!(recording_hook.after_result.lock().unwrap().is_none());
-        assert!(!db.cloudsync_interrupt_registered());
-        tokio::time::timeout(
-            Duration::from_millis(250),
-            db.cloudsync_wait_for_sync_idle(),
-        )
-        .await
-        .expect("pending payload worker did not become idle after activity pause");
-        tokio::time::timeout(
-            Duration::from_millis(250),
-            sqlx::query("INSERT INTO items (id, value) VALUES ('after', 'local')")
-                .execute(db.pool()),
-        )
-        .await
-        .expect("immediate local write remained blocked after activity pause")
-        .unwrap();
-    }
-
-    #[tokio::test]
-    async fn paused_manual_sync_preserves_the_last_result_and_error() {
-        let db = Db::connect_memory().await.unwrap();
-        let hook = Arc::new(RecordingSyncHook {
-            activity_paused: AtomicBool::new(true),
-            ..Default::default()
-        });
-        db.set_cloudsync_sync_hook(hook);
-        db.cloudsync_configure(test_cloudsync_config())
-            .await
-            .unwrap();
-        let last_sync = CloudsyncNetworkResult {
-            send: None,
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 3,
-                tables: vec!["sessions".to_string()],
-                chunks: 1,
-                bytes: 2048,
-                complete: true,
-                error: None,
-                last_failure: None,
-            }),
-        };
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.network_initialized = true;
-            runtime.last_sync = Some(last_sync.clone());
-            runtime.last_sync_at_ms = Some(42);
-            runtime.last_error = Some("previous error".to_string());
-            runtime.consecutive_failures = 2;
-        }
-
-        assert_eq!(
-            db.cloudsync_trigger_sync().await.unwrap(),
-            CloudsyncNetworkResult::default()
-        );
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(status.activity_paused);
-        assert_eq!(status.has_unsent_changes, None);
-        assert_eq!(status.last_sync.as_ref(), Some(&last_sync));
-        assert_eq!(status.last_sync_at_ms, Some(42));
-        assert_eq!(status.last_error.as_deref(), Some("previous error"));
-        assert_eq!(status.consecutive_failures, 2);
-    }
-
-    #[tokio::test]
-    async fn sync_idle_barrier_waits_for_the_active_operation() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        let operation = db.cloudsync_sync_operation.lock().await;
-        let mut barrier = Box::pin(db.cloudsync_wait_for_sync_idle());
-
-        tokio::select! {
-            biased;
-            () = &mut barrier => panic!("sync idle barrier bypassed the active operation"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        drop(operation);
-        tokio::time::timeout(Duration::from_millis(100), barrier)
-            .await
-            .expect("sync idle barrier did not finish");
-    }
-
-    #[tokio::test]
-    async fn requested_sync_wakes_are_coalesced() {
-        let db = Db::connect_memory_plain().await.unwrap();
-
-        db.cloudsync_request_sync();
-        assert!(
-            tokio::time::timeout(
-                Duration::from_millis(10),
-                db.cloudsync_sync_requested.notified(),
-            )
-            .await
-            .is_err(),
-            "a stopped runtime retained a requested sync wake"
-        );
-        db.cloudsync_runtime.lock().unwrap().running = true;
-        db.cloudsync_request_sync();
-        db.cloudsync_request_sync();
-        tokio::time::timeout(
-            Duration::from_millis(100),
-            db.cloudsync_sync_requested.notified(),
-        )
-        .await
-        .expect("queued sync wake was lost");
-        assert!(
-            tokio::time::timeout(
-                Duration::from_millis(10),
-                db.cloudsync_sync_requested.notified(),
-            )
-            .await
-            .is_err(),
-            "duplicate sync wakes were not coalesced"
-        );
-    }
-
-    #[tokio::test]
-    async fn requested_sync_interrupts_and_consumes_the_retry_backoff() {
-        let sync_requested = tokio::sync::Notify::new();
-        let (_shutdown_tx, mut shutdown_rx) = tokio::sync::oneshot::channel();
-
-        sync_requested.notify_one();
-        sync_requested.notify_one();
-        tokio::time::timeout(
-            Duration::from_millis(100),
-            wait_for_retry_request_or_shutdown(
-                Duration::from_secs(60),
-                &sync_requested,
-                &mut shutdown_rx,
-            ),
-        )
-        .await
-        .expect("requested sync did not interrupt retry backoff");
-        assert!(
-            tokio::time::timeout(
-                Duration::from_millis(10),
-                wait_for_retry_request_or_shutdown(
-                    Duration::from_secs(60),
-                    &sync_requested,
-                    &mut shutdown_rx,
-                ),
-            )
-            .await
-            .is_err(),
-            "the retry wake was left queued for a duplicate sync"
-        );
-    }
-
-    #[test]
-    fn requested_sync_retries_promptly_when_another_sync_is_busy() {
-        let interval = Duration::from_secs(30);
-        let request_pending = cloudsync_request_pending(false, CloudsyncWake::Requested);
-        let request_pending = cloudsync_request_pending(request_pending, CloudsyncWake::Interval);
-
-        assert_eq!(
-            cloudsync_busy_delay(request_pending, interval),
-            CLOUDSYNC_PROGRESS_INTERVAL,
-            "a timer collision must preserve the pending requested sync"
-        );
-        assert_eq!(cloudsync_busy_delay(false, interval), interval);
-    }
-
-    #[tokio::test]
-    async fn after_sync_hook_receives_the_bounded_network_result() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        let expected = CloudsyncNetworkResult {
-            send: Some(anlg_cloudsync::NetworkSendResult {
-                status: "synced".to_string(),
-                local_version: 4,
-                server_version: 4,
-                chunks: 1,
-                bytes: 1024,
-                last_failure: None,
-            }),
-            receive: Some(anlg_cloudsync::NetworkReceiveResult {
-                rows: 3,
-                tables: vec!["sessions".to_string()],
-                chunks: 1,
-                bytes: 2048,
-                complete: false,
-                error: None,
-                last_failure: None,
-            }),
-        };
-        let recording_hook = Arc::new(RecordingSyncHook::default());
-        let hook: Arc<dyn crate::CloudsyncSyncHook> = recording_hook.clone();
-        let hook = Mutex::new(Some(hook));
-
-        run_after_sync_hook(&hook, db.pool(), &expected)
-            .await
-            .unwrap();
-
-        assert_eq!(
-            recording_hook.after_result.lock().unwrap().as_ref(),
-            Some(&expected)
-        );
-    }
-
-    #[tokio::test]
-    async fn logout_releases_connection_after_partial_startup() {
-        let mut db = Db::connect_memory_plain().await.unwrap();
-        db.cloudsync_enabled = true;
-        db.cloudsync_configure(test_cloudsync_config())
-            .await
-            .unwrap();
-        *db.cloudsync_connection.lock().await = Some(db.pool.acquire().await.unwrap());
-        db.cloudsync_runtime.lock().unwrap().outbound_work_state = Some(false);
-
-        db.cloudsync_logout(false).await.unwrap();
-
-        assert!(db.cloudsync_connection.lock().await.is_none());
-        let runtime = db.cloudsync_runtime.lock().unwrap();
-        assert!(runtime.config.is_none());
-        assert!(runtime.outbound_work_state.is_none());
-    }
-
-    #[tokio::test]
-    async fn logout_signals_shutdown_before_waiting_for_an_active_sync_operation() {
-        let mut db = Db::connect_memory_plain().await.unwrap();
-        db.cloudsync_enabled = true;
-        db.cloudsync_configure(test_cloudsync_config())
-            .await
-            .unwrap();
-        let db = Arc::new(db);
-        let (shutdown_tx, shutdown_rx) = oneshot::channel();
-        let (shutdown_observed_tx, shutdown_observed_rx) = oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            shutdown_rx.await.unwrap();
-            let _ = shutdown_observed_tx.send(());
-        });
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.running = true;
-            runtime.task = Some(CloudsyncBackgroundTask {
-                shutdown_tx: Some(shutdown_tx),
-                join_handle,
-            });
-        }
-        let sync_operation = db.cloudsync_sync_operation.lock().await;
-        let logout_db = Arc::clone(&db);
-        let logout = tokio::spawn(async move { logout_db.cloudsync_logout(false).await });
-
-        tokio::time::timeout(Duration::from_secs(1), shutdown_observed_rx)
-            .await
-            .expect("logout waited for the sync operation before signaling shutdown")
-            .unwrap();
-        assert!(!logout.is_finished());
-
-        drop(sync_operation);
-        logout.await.unwrap().unwrap();
-    }
-
-    #[tokio::test]
-    async fn task_only_resume_reuses_initialized_transport_and_preserves_error_state() {
-        let mut db = Db::connect_memory_plain().await.unwrap();
-        db.cloudsync_enabled = true;
-        let config = test_cloudsync_config();
-        db.cloudsync_configure(config.clone()).await.unwrap();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.network_initialized = true;
-            runtime.last_error = Some("existing sync failure".to_string());
-            runtime.consecutive_failures = 2;
-        }
-        let sync_operation = db.cloudsync_sync_operation.lock().await;
-
-        db.start_cloudsync_background_task(&config);
-
-        {
-            let runtime = db.cloudsync_runtime.lock().unwrap();
-            assert!(runtime.running);
-            assert!(runtime.network_initialized);
-            assert!(runtime.task.is_some());
-            assert_eq!(runtime.last_error.as_deref(), Some("existing sync failure"));
-            assert_eq!(runtime.consecutive_failures, 2);
-        }
-        db.stop_cloudsync_task().await;
-        drop(sync_operation);
-    }
-
-    #[tokio::test]
-    async fn authentication_failure_cleans_up_initialized_network() {
-        let cleanup_called = AtomicBool::new(false);
-
-        let error = authenticate_cloudsync_network(
-            || async {
-                Err::<(), _>(anlg_cloudsync::Error::from(std::io::Error::other(
-                    "authentication rejected",
-                )))
-            },
-            || async {
-                cleanup_called.store(true, Ordering::SeqCst);
-                Ok::<(), anlg_cloudsync::Error>(())
-            },
-        )
-        .await
-        .unwrap_err();
-
-        assert!(cleanup_called.load(Ordering::SeqCst));
-        assert!(error.to_string().contains("authentication rejected"));
-    }
-
-    #[tokio::test]
-    async fn manual_transport_initializes_without_starting_background_sync() {
-        let db = Db::connect_memory().await.unwrap();
-        sqlx::query("CREATE TABLE items (id TEXT PRIMARY KEY NOT NULL)")
-            .execute(db.pool())
-            .await
-            .unwrap();
-        let mut config = test_cloudsync_config();
-        config.tables = vec![CloudsyncTableSpec {
-            table_name: "items".to_string(),
-            crdt_algo: None,
-            init_flags: None,
-            enabled: true,
-        }];
-
-        db.cloudsync_prepare_manual_transport(config).await.unwrap();
-
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(status.configured);
-        assert!(!status.running);
-        assert!(status.network_initialized);
-        assert!(db.cloudsync_runtime.lock().unwrap().task.is_none());
-        let pending = db.cloudsync_manual_pending_payload_batch().await.unwrap();
-        assert_eq!(pending.chunks, 0);
-        assert!(pending.complete);
-        assert!(pending.fits);
-        db.cloudsync_suspend().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn manual_transport_refuses_to_take_over_a_running_runtime() {
-        let db = Db::connect_memory().await.unwrap();
-        db.cloudsync_runtime.lock().unwrap().running = true;
-
-        let error = db
-            .cloudsync_prepare_manual_transport(test_cloudsync_config())
-            .await
-            .unwrap_err();
-
-        assert!(matches!(error, CloudsyncRuntimeError::RestartRequired));
-        db.cloudsync_runtime.lock().unwrap().running = false;
-    }
-
-    #[tokio::test]
-    async fn prepared_transport_resumes_without_reinitializing_the_connection() {
-        let dir = tempfile::tempdir().unwrap();
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&dir.path().join("app.db")),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(2),
-        })
-        .await
-        .unwrap();
-        let config = test_cloudsync_config();
-        db.cloudsync_prepare_manual_transport(config).await.unwrap();
-        {
-            let mut connection = db.cloudsync_connection.lock().await;
-            sqlx::query("CREATE TEMP TABLE transport_marker (value INTEGER)")
-                .execute(&mut **connection.as_mut().unwrap())
-                .await
-                .unwrap();
-        }
-        let sync_operation = db.cloudsync_sync_operation.lock().await;
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.last_error = Some("stale recovery error".to_string());
-            runtime.last_error_kind = Some(anlg_cloudsync::ErrorKind::Transient);
-            runtime.consecutive_failures = 3;
-            runtime.outbound_work_state = Some(true);
-        }
-
-        db.cloudsync_resume_prepared_transport().await.unwrap();
-
-        let marker_exists: i64 = {
-            let mut connection = db.cloudsync_connection.lock().await;
-            sqlx::query_scalar(
-                "SELECT COUNT(*) FROM sqlite_temp_master WHERE name = 'transport_marker'",
-            )
-            .fetch_one(&mut **connection.as_mut().unwrap())
-            .await
-            .unwrap()
-        };
-        assert_eq!(marker_exists, 1);
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(status.running);
-        assert!(status.network_initialized);
-        assert!(status.last_error.is_none());
-        assert!(status.last_error_kind.is_none());
-        assert_eq!(status.consecutive_failures, 0);
-        assert_eq!(status.has_unsent_changes, None);
-
-        db.stop_cloudsync_task().await;
-        drop(sync_operation);
-        db.cloudsync_suspend().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn prepared_transport_reaps_a_finished_background_task_before_resuming() {
-        let dir = tempfile::tempdir().unwrap();
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Local(&dir.path().join("app.db")),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(2),
-        })
-        .await
-        .unwrap();
-        db.cloudsync_prepare_manual_transport(test_cloudsync_config())
-            .await
-            .unwrap();
-        let sync_operation = db.cloudsync_sync_operation.lock().await;
-        let join_handle = tokio::spawn(async {});
-        while !join_handle.is_finished() {
-            tokio::task::yield_now().await;
-        }
-        let (shutdown_tx, _shutdown_rx) = oneshot::channel();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.running = false;
-            runtime.task = Some(CloudsyncBackgroundTask {
-                shutdown_tx: Some(shutdown_tx),
-                join_handle,
-            });
-        }
-
-        db.cloudsync_resume_prepared_transport().await.unwrap();
-
-        {
-            let runtime = db.cloudsync_runtime.lock().unwrap();
-            assert!(runtime.running);
-            assert!(runtime.network_initialized);
-            assert!(runtime.task.is_some());
-        }
-
-        db.stop_cloudsync_task().await;
-        drop(sync_operation);
-        db.cloudsync_suspend().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn configure_rejects_mutation_while_manual_transport_is_prepared() {
-        let db = Db::connect_memory().await.unwrap();
-        let original = test_cloudsync_config();
-        db.cloudsync_prepare_manual_transport(original.clone())
-            .await
-            .unwrap();
-        let mut replacement = original.clone();
-        replacement.connection_string = "replacement-managed-database-id".to_string();
-
-        let error = db.cloudsync_configure(replacement).await.unwrap_err();
-
-        assert!(matches!(error, CloudsyncRuntimeError::RestartRequired));
-        assert_eq!(
-            db.cloudsync_runtime.lock().unwrap().config.as_ref(),
-            Some(&original)
-        );
-        db.cloudsync_suspend().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn reconfigure_cleans_up_a_prepared_manual_transport() {
-        let db = Db::connect_memory().await.unwrap();
-        db.cloudsync_prepare_manual_transport(test_cloudsync_config())
-            .await
-            .unwrap();
-        let mut replacement = test_cloudsync_config();
-        replacement.connection_string = "replacement-managed-database-id".to_string();
-
-        db.cloudsync_reconfigure(replacement.clone()).await.unwrap();
-
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(status.configured);
-        assert!(!status.running);
-        assert!(!status.network_initialized);
-        assert_eq!(
-            db.cloudsync_runtime.lock().unwrap().config.as_ref(),
-            Some(&replacement)
-        );
-        db.cloudsync_suspend().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn configure_start_and_suspend_transitions_are_serialized() {
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Memory,
-            cloudsync_enabled: false,
-            journal_mode_wal: false,
-            foreign_keys: true,
-            max_connections: Some(1),
-        })
-        .await
-        .unwrap();
-        db.cloudsync_configure(CloudsyncRuntimeConfig {
-            connection_string: "managed-database-id".to_string(),
-            auth: CloudsyncAuth::None,
-            tables: Vec::new(),
-            sync_interval_ms: 30_000,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        })
-        .await
-        .unwrap();
-
-        let lifecycle = db.cloudsync_lifecycle.lock().await;
-        let mut configure = Box::pin(db.cloudsync_configure(CloudsyncRuntimeConfig {
-            connection_string: "next-managed-database-id".to_string(),
-            auth: CloudsyncAuth::None,
-            tables: Vec::new(),
-            sync_interval_ms: 45_000,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        }));
-        tokio::select! {
-            biased;
-            result = &mut configure => panic!("configure bypassed lifecycle lock: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        let mut start = Box::pin(db.cloudsync_start());
-        tokio::select! {
-            biased;
-            result = &mut start => panic!("start bypassed lifecycle lock: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        let mut suspend = Box::pin(db.cloudsync_suspend());
-        tokio::select! {
-            biased;
-            result = &mut suspend => panic!("suspend bypassed lifecycle lock: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        drop(lifecycle);
-        configure.await.unwrap();
-        start.await.unwrap();
-        suspend.await.unwrap();
-
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(!status.configured);
-        assert!(!status.running);
-        assert!(!status.network_initialized);
-    }
-
-    #[tokio::test]
-    async fn status_stays_observable_while_mutations_wait_for_lifecycle() {
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Memory,
-            cloudsync_enabled: false,
-            journal_mode_wal: false,
-            foreign_keys: true,
-            max_connections: Some(1),
-        })
-        .await
-        .unwrap();
-        db.cloudsync_configure(CloudsyncRuntimeConfig {
-            connection_string: "managed-database-id".to_string(),
-            auth: CloudsyncAuth::None,
-            tables: Vec::new(),
-            sync_interval_ms: 30_000,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        })
-        .await
-        .unwrap();
-
-        let lifecycle = db.cloudsync_lifecycle.lock().await;
-        let mut suspend = Box::pin(db.cloudsync_suspend());
-        tokio::select! {
-            biased;
-            result = &mut suspend => panic!("suspend bypassed lifecycle lock: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        let status = tokio::time::timeout(Duration::from_millis(100), db.cloudsync_status())
-            .await
-            .expect("status blocked on the lifecycle lock")
-            .unwrap();
-        assert!(status.configured);
-        assert!(!status.running);
-        assert_eq!(status.has_unsent_changes, None);
-
-        let mut trigger = Box::pin(db.cloudsync_trigger_sync());
-        tokio::select! {
-            biased;
-            result = &mut trigger => panic!("manual sync bypassed lifecycle lock: {result:?}"),
-            _ = tokio::task::yield_now() => {}
-        }
-
-        drop(lifecycle);
-        suspend.await.unwrap();
-        assert_eq!(trigger.await.unwrap(), CloudsyncNetworkResult::default());
-        assert!(!db.cloudsync_status().await.unwrap().configured);
-    }
-
-    #[tokio::test]
-    async fn status_preserves_unknown_outbound_state_during_sync_preflight() {
-        let db = Db::connect_memory().await.unwrap();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.running = true;
-            runtime.network_initialized = true;
-        }
-        let _sync_operation = db.cloudsync_sync_operation.lock().await;
-
-        let status = tokio::time::timeout(Duration::from_millis(100), db.cloudsync_status())
-            .await
-            .expect("status blocked on the active sync operation")
-            .unwrap();
-
-        assert!(status.configured);
-        assert!(status.running);
-        assert!(status.network_initialized);
-        assert_eq!(status.has_unsent_changes, None);
-    }
-
-    #[tokio::test]
-    async fn status_reports_receive_only_work_while_a_sync_operation_is_running() {
-        let db = Db::connect_memory().await.unwrap();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.running = true;
-            runtime.network_initialized = true;
-            runtime.last_sync_at_ms = Some(42);
-            runtime.outbound_work_state = Some(false);
-        }
-        let _sync_operation = db.cloudsync_sync_operation.lock().await;
-
-        let status = tokio::time::timeout(Duration::from_millis(100), db.cloudsync_status())
-            .await
-            .expect("status blocked on the active receive")
-            .unwrap();
-
-        assert_eq!(status.has_unsent_changes, Some(false));
-        assert_eq!(status.last_sync_at_ms, Some(42));
-    }
-
-    #[tokio::test]
-    async fn status_reports_outbound_work_while_a_sync_operation_is_running() {
-        let db = Db::connect_memory().await.unwrap();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.running = true;
-            runtime.network_initialized = true;
-            runtime.last_sync_at_ms = Some(42);
-            runtime.outbound_work_state = Some(true);
-        }
-        let _sync_operation = db.cloudsync_sync_operation.lock().await;
-
-        let status = tokio::time::timeout(Duration::from_millis(100), db.cloudsync_status())
-            .await
-            .expect("status blocked on the active send")
-            .unwrap();
-
-        assert_eq!(status.has_unsent_changes, Some(true));
-    }
-
-    #[tokio::test]
-    async fn status_reads_unsent_changes_without_network_io() {
-        let (_dir, db) = db_with_local_unsent_changes().await;
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.running = true;
-            runtime.network_initialized = true;
-        }
-
-        let status = tokio::time::timeout(Duration::from_millis(100), db.cloudsync_status())
-            .await
-            .expect("local cloudsync status blocked")
-            .unwrap();
-
-        assert_eq!(status.has_unsent_changes, Some(true));
-    }
-
-    #[tokio::test]
-    async fn repeated_status_polling_does_not_queue_work_on_a_busy_pool() {
-        let (_dir, db) = db_with_local_unsent_changes().await;
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.running = true;
-            runtime.network_initialized = true;
-        }
-        let mut first_connection = db.pool().acquire().await.unwrap();
-
-        for _ in 0..32 {
-            let status = tokio::time::timeout(Duration::from_millis(50), db.cloudsync_status())
-                .await
-                .expect("CloudSync status waited for a busy pool")
-                .unwrap();
-            assert_eq!(status.has_unsent_changes, None);
-        }
-
-        first_connection.return_to_pool().await;
-        for _ in 0..32 {
-            let status = tokio::time::timeout(Duration::from_millis(100), db.cloudsync_status())
-                .await
-                .expect("CloudSync status left SQLite work in flight")
-                .unwrap();
-            assert_eq!(status.has_unsent_changes, Some(true));
-        }
-        assert!(db.pool().num_idle() > 0);
-        tokio::time::timeout(Duration::from_millis(100), db.pool().acquire())
-            .await
-            .expect("repeated CloudSync status polling exhausted the pool")
-            .unwrap();
-    }
-
-    #[tokio::test]
-    async fn logout_checks_unsent_changes_without_network_io() {
-        let (_dir, db) = db_with_local_unsent_changes().await;
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.config = Some(test_cloudsync_config());
-            runtime.network_initialized = true;
-        }
-
-        let error = db.cloudsync_logout(false).await.unwrap_err();
-
-        assert!(
-            matches!(error, CloudsyncRuntimeError::UnsentChanges),
-            "{error:?}"
-        );
-    }
-
-    #[tokio::test]
-    async fn restart_after_fatal_exit_cleans_native_state() {
-        let tmp = tempfile::tempdir().unwrap();
-        let db_path = tmp.path().join("app.db");
-        let db = Db::open(crate::DbOpenOptions {
-            storage: crate::DbStorage::Local(&db_path),
-            cloudsync_enabled: true,
-            journal_mode_wal: true,
-            foreign_keys: true,
-            max_connections: Some(2),
-        })
-        .await
-        .unwrap();
-        db.cloudsync_configure(CloudsyncRuntimeConfig {
-            connection_string: "managed-database-id".to_string(),
-            auth: super::super::CloudsyncAuth::None,
-            tables: Vec::new(),
-            sync_interval_ms: 30_000,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        })
-        .await
-        .unwrap();
-        db.cloudsync_start().await.unwrap();
-        {
-            let mut connection = db.cloudsync_connection.lock().await;
-            sqlx::query("CREATE TEMP TABLE stale_cloudsync_connection (id INTEGER)")
-                .execute(&mut **connection.as_mut().unwrap())
-                .await
-                .unwrap();
-        }
-
-        let mut running_task = db.cloudsync_runtime.lock().unwrap().task.take().unwrap();
-        let _ = running_task.shutdown_tx.take().unwrap().send(());
-        let _ = running_task.join_handle.await;
-
-        let (stale_shutdown_tx, stale_shutdown_rx) = oneshot::channel::<()>();
-        let (finished_tx, finished_rx) = oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            drop(stale_shutdown_rx);
-            let _ = finished_tx.send(());
-        });
-        finished_rx.await.unwrap();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.running = false;
-            runtime.last_error = Some("fatal sync failure".to_string());
-            runtime.last_error_kind = Some(anlg_cloudsync::ErrorKind::Fatal);
-            runtime.task = Some(CloudsyncBackgroundTask {
-                shutdown_tx: Some(stale_shutdown_tx),
-                join_handle,
-            });
-        }
-
-        db.cloudsync_start().await.unwrap();
-
-        {
-            let runtime = db.cloudsync_runtime.lock().unwrap();
-            assert!(runtime.running);
-            assert!(runtime.network_initialized);
-            assert!(runtime.task.is_some());
-            assert!(runtime.last_error.is_none());
-        }
-        let marker_count: i64 = {
-            let mut connection = db.cloudsync_connection.lock().await;
-            sqlx::query_scalar(
-                "SELECT COUNT(*) FROM sqlite_temp_master WHERE name = 'stale_cloudsync_connection'",
-            )
-            .fetch_one(&mut **connection.as_mut().unwrap())
-            .await
-            .unwrap()
-        };
-        assert_eq!(marker_count, 0);
-        db.cloudsync_stop().await.unwrap();
-    }
-
-    #[tokio::test]
-    async fn suspend_interrupts_active_retry_backoff() {
-        let db = Db::connect_memory_plain().await.unwrap();
-        let (shutdown_tx, mut shutdown_rx) = oneshot::channel();
-        let (retry_started_tx, retry_started_rx) = oneshot::channel();
-        let join_handle = tokio::spawn(async move {
-            let sync_requested = tokio::sync::Notify::new();
-            let _ = retry_started_tx.send(());
-            assert!(
-                !wait_for_retry_request_or_shutdown(
-                    Duration::from_secs(60),
-                    &sync_requested,
-                    &mut shutdown_rx,
-                )
-                .await
-            );
-        });
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.running = true;
-            runtime.task = Some(CloudsyncBackgroundTask {
-                shutdown_tx: Some(shutdown_tx),
-                join_handle,
-            });
-        }
-        retry_started_rx.await.unwrap();
-
-        tokio::time::timeout(Duration::from_secs(1), db.cloudsync_suspend())
-            .await
-            .expect("suspend waited for retry backoff")
-            .unwrap();
-
-        assert!(!db.cloudsync_status().await.unwrap().running);
-    }
-
-    #[tokio::test]
-    async fn shutdown_interrupts_an_active_sync_future() {
-        let (shutdown_tx, mut shutdown_rx) = oneshot::channel();
-        shutdown_tx.send(()).unwrap();
-
-        let result = tokio::time::timeout(
-            Duration::from_secs(1),
-            run_or_shutdown(pending::<()>(), &mut shutdown_rx),
-        )
-        .await
-        .expect("active sync ignored shutdown");
-
-        assert!(result.is_none());
-    }
-
-    #[tokio::test]
-    async fn suspend_stops_runtime_and_clears_in_memory_credentials() {
-        let db = Db::open(DbOpenOptions {
-            storage: DbStorage::Memory,
-            cloudsync_enabled: false,
-            journal_mode_wal: false,
-            foreign_keys: true,
-            max_connections: Some(1),
-        })
-        .await
-        .unwrap();
-        db.cloudsync_configure(CloudsyncRuntimeConfig {
-            connection_string: "managed-database-id".to_string(),
-            auth: CloudsyncAuth::Token {
-                token: "secret-token".to_string(),
-            },
-            tables: vec![CloudsyncTableSpec {
-                table_name: "sessions".to_string(),
-                crdt_algo: None,
-                init_flags: None,
-                enabled: true,
-            }],
-            sync_interval_ms: 30_000,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        })
-        .await
-        .unwrap();
-
-        db.cloudsync_start().await.unwrap();
-        db.cloudsync_runtime.lock().unwrap().outbound_work_state = Some(false);
-        db.cloudsync_suspend().await.unwrap();
-
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(!status.configured);
-        assert!(!status.running);
-        assert!(!status.network_initialized);
-        assert!(
-            db.cloudsync_runtime
-                .lock()
-                .unwrap()
-                .outbound_work_state
-                .is_none()
-        );
-    }
-
-    #[tokio::test]
-    async fn suspend_clears_runtime_state_when_native_teardown_fails() {
-        let db = Db::connect_memory().await.unwrap();
-        db.cloudsync_configure(CloudsyncRuntimeConfig {
-            connection_string: "managed-database-id".to_string(),
-            auth: CloudsyncAuth::Token {
-                token: "secret-token".to_string(),
-            },
-            tables: Vec::new(),
-            sync_interval_ms: 30_000,
-            wait_ms: Some(5_000),
-            max_retries: Some(3),
-        })
-        .await
-        .unwrap();
-        {
-            let mut runtime = db.cloudsync_runtime.lock().unwrap();
-            runtime.running = true;
-            runtime.network_initialized = true;
-        }
-        db.pool().close().await;
-
-        db.cloudsync_suspend().await.unwrap_err();
-
-        let status = db.cloudsync_status().await.unwrap();
-        assert!(!status.configured);
-        assert!(!status.running);
-        assert!(!status.network_initialized);
-        assert!(db.cloudsync_connection.lock().await.is_none());
-    }
-}
+mod tests;
 
 fn record_sync_error(runtime: &Mutex<CloudsyncRuntimeState>, error: &anlg_cloudsync::Error) {
     let mut runtime = runtime.lock().unwrap();
